Broke addClickFunctions out as SubmissionClickActions.java from PopulateSubmissionViewHolder.java
